Move some file processing functions into their own modules.
